WIN、LOSEのビューを共通化した
